Testing and NA treatment
In helper_funtions I changed the is.symmetric function to be able to state
that D matrices are symmetric even if they have NAs (as long as those are at
symmetric positions).

The input check was changed slightly: i.e. if a distance matrix has all zero
entries, the cmds function throws an error.